Meter unitIndex test 2 (unsigned int)

+++ Meter.c
@@ -54,7 +54,7 @@ int Meter_humanUnit(char* buffer, unsigned long int value, size_t size) {
    const char prefix[] = {'K','M','G','T','P','E','Z','Y'};
    unsigned long int powi = 1;
    unsigned int powj = 1, precision = 2;
-   size_t unitIndex = 0;
+   unsigned int unitIndex = 0;
 
    for (;;) {
       if (value / 1024 < powi)
